refactor: code template
