+++ src/api.ts
@@ -1,6 +1,5 @@
 export * as TypeScriptCodeGenerator from "./internal/CodeGenerator";
 export * as Converter from "./internal/Converter";
-export * as DefaultCodeTemplate from "./DefaultCodeTemplate";
 export { fileSystem } from "./internal/FileSystem";
 export * as ResolveReference from "./internal/ResolveReference";
 export * as Validator from "./internal/Validator";
