+++ src/code-templates/api-client/ApiClientArgument.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 
-import type { Factory } from "../factory";
-import type { CodeGenerator } from "../types";
+import type { Factory } from "../../factory";
+import type { CodeGenerator } from "../../types";
 
 /**
  * export type RequestContentType${operationId} = keyof RequestBody${operationId};
