+++ src/code-templates/api-client/ApiClientClass/ApiClientInterface.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 
-import type { Factory } from "../../factory";
-import type { CodeGenerator } from "../../types";
+import type { Factory } from "../../../factory";
+import type { CodeGenerator } from "../../../types";
 
 const httpMethodList: string[] = ["GET", "PUT", "POST", "DELETE", "OPTIONS", "HEAD", "PATCH", "TRACE"];
