+++ src/code-templates/api-client/ApiClientClass/Method.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 
-import type { Factory } from "../../factory";
-import type { CodeGenerator } from "../../types";
+import type { Factory } from "../../../factory";
+import type { CodeGenerator } from "../../../types";
 import * as MethodBody from "./MethodBody";
 
 export { MethodBody };
