+++ src/code-templates/api-client/ApiClientClass/MethodBody/index.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 
-import type { Factory } from "../../../factory";
-import type { CodeGenerator } from "../../../types";
+import type { Factory } from "../../../../factory";
+import type { CodeGenerator } from "../../../../types";
 import * as Utils from "../../utils";
 import * as CallRequest from "./CallRequest";
 import * as HeaderParameter from "./HeaderParameter";
