+++ src/templates.ts
@@ -1,0 +1,1 @@
+export * as ApiClient from "./code-templates/api-client";
